// api/model/submissionDto.ts
/**
 * Student-Management-System-API
 * The Student-Management-System-API. <a href=\'http://localhost:3000/api-json\'>JSON</a>
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LinkDto } from './linkDto';


export interface SubmissionDto { 
    assignmentId: string;
    userId: string;
    displayName: string;
    date: string;
    groupId?: string;
    groupName?: string;
    links?: Array<LinkDto>;
    payload?: any;
}